* Removing the `panoptes.utils.data.assets` module and moving the `Downloader` class into the `scripts/download-data.py` file directly.
